Modify weekly swiper
